chat_server/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
boto3==1.34.122
botocore==1.34.122
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.6
distro==1.9.0
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.3
langchain-openai==0.0.2
langchain-text-splitters==0.2.0
langchainhub==0.1.14
langserve==0.2.0
langsmith==0.1.67
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.7.2
orjson==3.10.3
packaging==23.2
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
pyproject-toml==0.0.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.0
rich==13.7.1
rpds-py==0.18.1
s3transfer==0.10.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
sse-starlette==1.8.2
starlette==0.37.2
tenacity==8.3.0
tiktoken==0.5.2
toml==0.10.2
tqdm==4.66.4
typer==0.12.3
types-requests==2.32.0.20240602
typing-inspect==0.9.0
typing_extensions==4.11.0
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
yarl==1.9.4
